app-server/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
fastapi==0.115.5
flatbuffers==24.3.25
h11==0.14.0
humanfriendly==10.0
idna==3.10
imageio==2.36.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.43.0
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
onnxruntime==1.20.0
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
PyMatting==1.1.13
python-multipart==0.0.17
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
rpds-py==0.21.0
scikit-image==0.24.0
scipy==1.14.1
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.3
tifffile==2024.9.20
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
